requirements.txt
amqp==2.5.2
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.7
autopep8==1.5.3
Babel==2.8.0
beautifulsoup4==4.7.1
billiard==3.6.3.0
boto3==1.13.11
botocore==1.16.11
braintree==4.1.0
cachetools==4.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cssselect2==0.3.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
dj-email-url==1.0.0
Django==3.0.6
django-address==0.2.2
django-appconf==1.0.4
django-cache-url==3.1.2
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.2.0
django-filters==0.2.1
django-graphql-jwt==0.3.0
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-render-block==0.6
django-rest-knox==4.1.0
django-storages==1.9.1
django-templated-email==2.3.0
django-versatileimagefield==2.0
django-weekday-field==1.1.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
docutils==0.15.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
Faker==4.1.0
filelock==3.0.12
freezegun==0.3.15
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.10.0
graphene-federation==0.1.0
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.9
install==1.3.3
jaeger-client==4.3.0
jmespath==0.9.5
jsonfield==3.1.0
kombu==4.6.8
lxml==4.5.0
Markdown==3.2.2
maxminddb==1.5.4
maxminddb-geolite2==2018.703
measurement==3.2.0
mpmath==1.1.0
numpy==1.18.5
oauthlib==3.1.0
opentracing==2.3.0
pandas==1.0.4
pathlib==1.0.1
phonenumberslite==8.12.3
Pillow==7.1.2
prices==1.0.0
promise==2.3
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.3
python-magic==0.4.18
python-magic-bin==0.4.14
pytz==2020.1
razorpay==1.2.0
redis==3.5.1
requests==2.23.0
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
sentry-sdk==0.14.4
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.48.0
sympy==1.5.1
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
tqdm==4.46.0
typing==3.7.4.1
Unipath==1.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.17
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
